+++ cl/phase1/core/state/ssz.go
@@ -1,29 +1,28 @@
 package state
 
 import (
-	"github.com/ledgerwatch/erigon/metrics/methelp"
-
 	"github.com/ledgerwatch/erigon-lib/types/clonable"
+	"github.com/ledgerwatch/erigon/metrics"
 )
 
 func (b *CachingBeaconState) EncodeSSZ(buf []byte) ([]byte, error) {
-	h := methelp.NewHistTimer("encode_ssz_beacon_state_dur")
+	h := metrics.NewHistTimer("encode_ssz_beacon_state_dur")
 	bts, err := b.BeaconState.EncodeSSZ(buf)
 	if err != nil {
 		return nil, err
 	}
 	h.PutSince()
-	sz := methelp.NewHistTimer("encode_ssz_beacon_state_size")
+	sz := metrics.NewHistTimer("encode_ssz_beacon_state_size")
 	sz.Update(float64(len(bts)))
 	return bts, err
 }
 
 func (b *CachingBeaconState) DecodeSSZ(buf []byte, version int) error {
-	h := methelp.NewHistTimer("decode_ssz_beacon_state_dur")
+	h := metrics.NewHistTimer("decode_ssz_beacon_state_dur")
 	if err := b.BeaconState.DecodeSSZ(buf, version); err != nil {
 		return err
 	}
-	sz := methelp.NewHistTimer("decode_ssz_beacon_state_size")
+	sz := metrics.NewHistTimer("decode_ssz_beacon_state_size")
 	sz.Update(float64(len(buf)))
 	h.PutSince()
 	return b.initBeaconState()
