+++ cmd/caplin-regression/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"flag"
 
-	"github.com/ledgerwatch/erigon/metrics/exp"
+	"github.com/ledgerwatch/erigon/metrics"
 	"github.com/ledgerwatch/erigon/turbo/debug"
 
 	"github.com/ledgerwatch/erigon/cl/cltypes"
@@ -43,7 +43,7 @@ func main() {
 	)
 	if *pprof {
 		// Server for pprof
-		debug.StartPProf("localhost:6060", exp.Setup("localhost:6060", log.Root()))
+		debug.StartPProf("localhost:6060", metrics.Setup("localhost:6060", log.Root()))
 	}
 
 	if err != nil {
