+++ cmd/erigon/main.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 
 	"github.com/ledgerwatch/erigon-lib/common/dbg"
-	metrics "github.com/ledgerwatch/erigon/metrics/methelp"
+	"github.com/ledgerwatch/erigon/metrics"
 	"github.com/ledgerwatch/log/v3"
 	"github.com/pelletier/go-toml"
 	"github.com/urfave/cli/v2"
